requirements.txt
# WWAI-GNN Requirements
# Graph Neural Network Economic Forecasting

# Core JAX ecosystem
jax>=0.4.20
jaxlib>=0.4.20
dm-haiku>=0.0.10
jraph>=0.0.6.dev0
optax>=0.1.7
chex>=0.1.82

# Data handling
numpy>=1.24.0
pandas>=2.0.0
xarray>=2023.10.0

# Data sources
fredapi>=0.5.0
wbgapi>=1.0.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# API Server
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Utilities
tqdm>=4.65.0
scipy>=1.10.0

# Testing (optional)
pytest>=7.0.0
pytest-cov>=4.0.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0
